+++ css/index.styl
@@ -1,7 +1,5 @@
 @import 'nib'
 
-global-reset()
-
 *
   box-sizing: border-box
 
@@ -39,6 +37,7 @@ body
     padding: .5rem
     font: inherit
     font-size: .875rem
+    line-height: 1
     color: #fff
     cursor: pointer
 
@@ -65,6 +64,6 @@ body
     transform: scale(.5)
     transform-origin: 100% 0
 
-@media (min-width: 800px)
+@media (min-width: 640px)
   html
     font-size: 125%
