+++ lib/card.js
@@ -1,13 +1,9 @@
 
+import Deck from './deck'
+
 import createElement from './create-element'
 import translate from './translate'
 import prefix from './prefix'
-import intro from './card-intro'
-import shuffle from './card-shuffle'
-import sort from './card-sort'
-import bysuit from './card-bysuit'
-import fan from './card-fan'
-import poker from './card-poker'
 
 var maxZ = 52
 
@@ -30,6 +26,9 @@ export default function (i) {
 
   var self = {i, value, suit, pos: i, $el, mount, unmount}
 
+  var modules = Deck.modules
+  var module
+
   $el.classList.add('card', suitName, (suitName + value))
   $topleft.classList.add('topleft')
   $bottomright.classList.add('bottomright')
@@ -45,12 +44,9 @@ export default function (i) {
   $el.appendChild($topleft)
   $el.appendChild($bottomright)
 
-  intro(self, $el)
-  shuffle(self, $el)
-  sort(self, $el)
-  bysuit(self, $el)
-  fan(self, $el)
-  poker(self, $el)
+  for (module in modules) {
+    addModule(modules[module])
+  }
 
   self.enableMoving = function () {
     if (isMovable) {
@@ -73,6 +69,10 @@ export default function (i) {
   }
 
   return self
+
+  function addModule (module) {
+    module.card && module.card(self)
+  }
 
   function onMousedown (e) {
     var middlePoint = self.$root.getBoundingClientRect()
