--- lib/deck-bysuit.js
@@ -1,16 +1,0 @@
-
-export default function (deck) {
-  deck.bysuit = deck.queued(bysuit)
-
-  function bysuit (next) {
-    var cards = deck.cards
-
-    cards.forEach(function (card) {
-      card.bysuit(function (i) {
-        if (i === cards.length - 1) {
-          next()
-        }
-      })
-    })
-  }
-}
