+++ lib/deck.js
@@ -1,12 +1,12 @@
 
 import createElement from './create-element'
 
-import shuffleModule from './deck-shuffle'
-import fanModule from './deck-fan'
-import sortModule from './deck-sort'
-import pokerModule from './deck-poker'
-import bysuitModule from './deck-bysuit'
-import introModule from './deck-intro'
+import bysuit from './modules/bysuit'
+import fan from './modules/fan'
+import intro from './modules/intro'
+import poker from './modules/poker'
+import shuffle from './modules/shuffle'
+import sort from './modules/sort'
 
 import observable from './observable'
 import queue from './queue'
@@ -23,13 +23,14 @@ export default function Deck (jokers) {
   var self = observable({mount, unmount, cards, $el})
   var $root
 
+  var modules = Deck.modules
+  var module
+
   queue(self)
-  shuffleModule(self)
-  sortModule(self)
-  fanModule(self)
-  pokerModule(self)
-  bysuitModule(self)
-  introModule(self)
+
+  for (module in modules) {
+    addModule(modules[module])
+  }
 
   $el.classList.add('deck')
 
@@ -50,7 +51,12 @@ export default function Deck (jokers) {
   function unmount () {
     $root.removeChild($el)
   }
+
+  function addModule (module) {
+    module.deck && module.deck(self)
+  }
 }
+Deck.modules = {bysuit, fan, intro, poker, shuffle, sort}
 Deck.Card = Card
 Deck.easing = easing
 Deck.prefix = prefix
