+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "start": "node index",
     "build-js": "rollup --name Deck -f iife lib/deck.js | babel -o dist/deck.js && uglifyjs dist/deck.js -cmo dist/deck.min.js",
-    "build-css": "stylus -u nib css/index.styl -c -o example/example.css",
+    "build-css": "stylus -u nib css/index.styl -o example/example.css",
     "build-html": "jade views/index.jade -o example/"
   },
   "repository": {
